Add files via upload
Figure 4. Comparison of the elements’ distributions throughout the weeks (1, 3, and 5) in the liver (Control and Tumor groups)
and primary tumor. 
Boxplots and violin plots in which light colors represent the liver, with light blue showing the Control group and
light red the Tumor group. The primary tumor is represented in dark red.